[Change] Simpler map visibility

+++ server/base/LocalClientHandlerFightManage.cpp
@@ -195,7 +195,7 @@ void Client::heal()
     rescue=unvalidated_rescue;
 }
 
-void Client::requestFight(const uint16_t &fightId)
+void Client::requestFight(const uint8_t &fightId)
 {
     if(isInFight())
     {
@@ -253,7 +253,7 @@ void Client::requestFight(const uint16_t &fightId)
     bool found=false;
     if(mapServer->botsFight.find(pos)!=mapServer->botsFight.cend())
     {
-        const std::vector<uint16_t> &botsFightList=mapServer->botsFight.at(pos);
+        const std::vector<uint8_t> &botsFightList=mapServer->botsFight.at(pos);
         if(vectorcontainsAtLeastOne(botsFightList,fightId))
             found=true;
     }
@@ -305,7 +305,7 @@ void Client::requestFight(const uint16_t &fightId)
         const std::pair<uint8_t,uint8_t> pos(x,y);
         if(mapServer->botsFight.find(pos)!=mapServer->botsFight.cend())
         {
-            const std::vector<uint16_t> &botsFightList=static_cast<MapServer*>(this->map)->botsFight.at(pos);
+            const std::vector<uint8_t> &botsFightList=static_cast<MapServer*>(this->map)->botsFight.at(pos);
             if(vectorcontainsAtLeastOne(botsFightList,fightId))
                 found=true;
         }
